requirements.txt
ccxt
numpy
pandas
ta
yfinance
anthropic
pydantic
httpx
python-dotenv
